# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pâtisserie ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /roles:
    get:
      operationId: getRoles
      tags: [roles]
      summary: List roles
      responses:
        "200":
          description: Roles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      tags: [roles]
      summary: Create role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleBody"
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /roles/{id}:
    get:
      operationId: getRole
      tags: [roles]
      summary: Get role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
    patch:
      operationId: updateRole
      tags: [roles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Update role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleBody"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"

  # USERS
...
  /recipes:
    get:
      operationId: getRecipes
      tags: [recipes]
      summary: List recipes
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Recipes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Create recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeBody"
      responses:
        "201":
          description: Recipe created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Get recipe
      responses:
        "200":
          description: Recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
    patch:
      operationId: updateRecipe
      tags: [recipes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Update recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecipeBody"
      responses:
        "200":
          description: Recipe updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"

  # PRODUCTION ORDERS
...
  /pos/sessions:
    get:
      operationId: getPOSSessions
      tags: [pos]
      summary: List POS sessions
      parameters:
        - name: branchId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: POS sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/POSSession"
    post:
      operationId: openPOSSession
      tags: [pos]
      summary: Open POS session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenPOSSessionBody"
      responses:
        "201":
          description: POS session opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/POSSession"
  /pos/sessions/{id}:
    get:
      operationId: getPOSSession
      tags: [pos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Get POS session
      responses:
        "200":
          description: POS session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/POSSession"
  /pos/sessions/{id}/close:
    post:
      operationId: closePOSSession
      tags: [pos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Close POS session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClosePOSSessionBody"
      responses:
        "200":
          description: POS session closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/POSSession"
...
components:
  schemas:
...
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        isSystem:
          type: boolean
        permissions:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - isSystem
        - permissions
        - createdAt
    CreateRoleBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        permissions:
          type: array
          items:
            type: string
      required:
        - name
        - permissions
    UpdateRoleBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        permissions:
          type: ["array", "null"]
          items:
            type: string
...
    Recipe:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        productId:
          type: ["integer", "null"]
        productName:
          type: ["string", "null"]
        type:
          type: string
          enum: [finished, semi_finished]
        yield:
          type: number
        yieldUnitId:
          type: integer
        yieldUnitName:
          type: string
        theoreticalCost:
          type: number
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        steps:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - yield
        - yieldUnitId
        - yieldUnitName
        - theoreticalCost
        - ingredients
        - createdAt
    RecipeIngredient:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        unitId:
          type: integer
        unitName:
          type: string
        wastageRate:
          type: number
        unitCost:
          type: number
        totalCost:
          type: number
      required:
        - id
        - productId
        - productName
        - quantity
        - unitId
        - unitName
        - wastageRate
        - unitCost
        - totalCost
    CreateRecipeBody:
      type: object
      properties:
        name:
          type: string
        productId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [finished, semi_finished]
        yield:
          type: number
        yieldUnitId:
          type: integer
        steps:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredientInput"
      required:
        - name
        - type
        - yield
        - yieldUnitId
        - ingredients
    UpdateRecipeBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        yield:
          type: ["number", "null"]
        steps:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        ingredients:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/RecipeIngredientInput"
    RecipeIngredientInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: number
        unitId:
          type: integer
        wastageRate:
          type: number
      required:
        - productId
        - quantity
        - unitId
        - wastageRate
...
    POSSession:
      type: object
      properties:
        id:
          type: integer
        branchId:
          type: integer
        branchName:
          type: string
        userId:
          type: integer
        userName:
          type: string
        status:
          type: string
          enum: [open, closed]
        openingCash:
          type: number
        countedCash:
          type: ["number", "null"]
        expectedCash:
          type: ["number", "null"]
        variance:
          type: ["number", "null"]
        totalSales:
          type: number
        totalCashSales:
          type: number
        totalCardSales:
          type: number
        salesCount:
          type: integer
        closureNotes:
          type: ["string", "null"]
        openedAt:
          type: string
        closedAt:
          type: ["string", "null"]
      required:
        - id
        - branchId
        - branchName
        - userId
        - userName
        - status
        - openingCash
        - totalSales
        - totalCashSales
        - totalCardSales
        - salesCount
        - openedAt
...
    OpenPOSSessionBody:
      type: object
      properties:
        branchId:
          type: integer
        openingCash:
          type: number
      required:
        - branchId
        - openingCash
    ClosePOSSessionBody:
      type: object
      properties:
        countedCash:
          type: number
        closureNotes:
          type: ["string", "null"]
      required:
        - countedCash
...